requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-auth==2.23.4
isodate==0.6.1
boto3==1.34.0

